WEBNEW-213 :truck: Moved render methods to a component dir
