+++ src/fifty-two/FiftyTwoPage.tsx
@@ -8,7 +8,7 @@ import { Category } from '../components/categoryFilter/CategoryFilter.types'
 import { Gallery } from '../components/gallery'
 import { Button } from '../components/button'
 import { GalleryCard } from '../components/galleryCard'
-import { renderMethods } from '../components/renderMethods'
+import { renderMethods } from '../components/richText/renderMethods'
 import { colors } from '../theme/colors'
 import { md } from '../theme/breakpoints'
 import { shuffle } from '../utils/iteration-utils'
